# apps/web/openapi/mediamop-openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MediaMop API",
    "version": "2.4.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/intake/webhook/{source_key}": {
      "post": {
        "description": "Accept one event from a media manager and hand it to whichever module owns it.",
        "operationId": "post_media_manager_intake_api_v1_intake_webhook__source_key__post",
        "parameters": [
          {
            "description": "Which media manager's payload dialect this body uses.",
            "in": "path",
            "name": "source_key",
            "required": true,
            "schema": {
              "description": "Which media manager's payload dialect this body uses.",
              "title": "Source Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Webhook-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Webhook-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Media Manager Intake Api V1 Intake Webhook  Source Key  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Media Manager Intake",
        "tags": [
          "media-manager-intake"
        ]
      }
    },
...
  }
}